+++ golang/lib/execution/test_suite_executor.go
@@ -1,29 +1,26 @@
-/*
- * Copyright (c) 2021 - present Kurtosis Technologies LLC.
- * All Rights Reserved.
- */
-
 package execution
 
 import (
-	"context"
+	"fmt"
 	"github.com/kurtosis-tech/kurtosis-libs/golang/lib/core_api_bindings"
-	"github.com/kurtosis-tech/kurtosis-libs/golang/lib/networks"
-	"github.com/kurtosis-tech/kurtosis-libs/golang/lib/testsuite"
+	"github.com/kurtosis-tech/kurtosis-libs/golang/lib/rpc_api/bindings"
+	"github.com/kurtosis-tech/kurtosis-libs/golang/lib/rpc_api/rpc_api_consts"
 	"github.com/palantir/stacktrace"
 	"github.com/sirupsen/logrus"
 	"google.golang.org/grpc"
-	"google.golang.org/protobuf/types/known/emptypb"
+	"net"
+	"os"
+	"os/signal"
+	"syscall"
 	"time"
 )
 
 const (
-	maxSuiteRegistrationRetries = 20
-	timeBetweenSuiteRegistrationRetries = 500 * time.Millisecond
+	grpcServerStopGracePeriod = 10 * time.Second
 )
 
 type TestSuiteExecutor struct {
-	kurtosisApiSocket string
+	kurtosisApiSocket string  // Can be empty if the testsuite is in metadata-providing mode
 	logLevelStr string
 	paramsJsonStr string
 	configurator TestSuiteConfigurator
@@ -33,7 +30,7 @@ func NewTestSuiteExecutor(kurtosisApiSocket string, logLevelStr string, paramsJs
 	return &TestSuiteExecutor{kurtosisApiSocket: kurtosisApiSocket, logLevelStr: logLevelStr, paramsJsonStr: paramsJsonStr, configurator: configurator}
 }
 
-func (executor *TestSuiteExecutor) Run(ctx context.Context) error {
+func (executor TestSuiteExecutor) Run() error {
 	if err := executor.configurator.SetLogLevel(executor.logLevelStr); err != nil {
 		return stacktrace.Propagate(err, "An error occurred setting the loglevel before running the testsuite")
 	}
@@ -43,146 +40,75 @@ func (executor *TestSuiteExecutor) Run(ctx context.Context) error {
 		return stacktrace.Propagate(err, "An error occurred parsing the suite params JSON and creating the testsuite")
 	}
 
-	// TODO SECURITY: Use HTTPS to ensure we're connecting to the real Kurtosis API servers
-	conn, err := grpc.Dial(executor.kurtosisApiSocket, grpc.WithInsecure())
+	var apiContainerService core_api_bindings.ApiContainerServiceClient = nil
+	if executor.kurtosisApiSocket != "" {
+		// TODO SECURITY: Use HTTPS to ensure we're connecting to the real Kurtosis API servers
+		conn, err := grpc.Dial(executor.kurtosisApiSocket, grpc.WithInsecure())
+		if err != nil {
+			return stacktrace.Propagate(
+				err,
+				"An error occurred creating a connection to the Kurtosis API server at '%v'",
+				executor.kurtosisApiSocket,
+			)
+		}
+		defer conn.Close()
+
+		apiContainerService = core_api_bindings.NewApiContainerServiceClient(conn)
+	}
+
+	testsuiteService := NewTestSuiteService(suite, apiContainerService)
+
+	// TODO all the code below here is almost entirely duplicated with KUrtosis Core - extract as a library!
+	grpcServer := grpc.NewServer()
+
+	bindings.RegisterTestSuiteServiceServer(grpcServer, testsuiteService)
+
+	listenAddressStr := fmt.Sprintf(":%v", rpc_api_consts.ListenPort)
+	listener, err := net.Listen(rpc_api_consts.ListenProtocol, listenAddressStr)
 	if err != nil {
 		return stacktrace.Propagate(
 			err,
-			"An error occurred creating a connection to the Kurtosis API server at '%v'",
-			executor.kurtosisApiSocket)
-	}
-	defer conn.Close()
-
-	suiteRegistrationClient := core_api_bindings.NewSuiteRegistrationServiceClient(conn)
-
-	var suiteRegistrationResp *core_api_bindings.SuiteRegistrationResponse
-	suiteRegistrationAttempts := 0
-	for {
-		if suiteRegistrationAttempts >= maxSuiteRegistrationRetries {
-			return stacktrace.NewError(
-				"Failed to register testsuite with API container, even after %v retries spaced %v apart",
-				maxSuiteRegistrationRetries,
-				timeBetweenSuiteRegistrationRetries)
-		}
-
-		resp, err := suiteRegistrationClient.RegisterSuite(ctx, &emptypb.Empty{})
-		if err == nil {
-			suiteRegistrationResp = resp
-			break
-		}
-		logrus.Debugf("The following error occurred registering testsuite with API container; retrying in %v: %v", timeBetweenSuiteRegistrationRetries, err)
-		time.Sleep(timeBetweenSuiteRegistrationRetries)
-		suiteRegistrationAttempts++
+			"An error occurred creating the listener on %v/%v",
+			rpc_api_consts.ListenProtocol,
+			listenAddressStr,
+		)
 	}
 
-	action := suiteRegistrationResp.SuiteAction
-	switch action {
-	case core_api_bindings.SuiteAction_SERIALIZE_SUITE_METADATA:
-		if err := runSerializeSuiteMetadataFlow(ctx, suite, conn); err != nil {
-			return stacktrace.Propagate(err, "An error occurred running the suite metadata serialization flow")
-		}
-		return nil
-	case core_api_bindings.SuiteAction_EXECUTE_TEST:
-		if err := runTestExecutionFlow(ctx, suite, conn); err != nil {
-			return stacktrace.Propagate(err, "An error occurred running the test execution flow")
-		}
-		return nil
-	default:
-		return stacktrace.NewError("Encountered unrecognized action '%v'; this is a bug in Kurtosis itself", action)
-	}
-}
+	// Docker will send SIGTERM to end the process, and we need to catch it to stop gracefully
+	termSignalChan := make(chan os.Signal, 1)
+	signal.Notify(termSignalChan, syscall.SIGINT, syscall.SIGTERM, syscall.SIGQUIT)
 
-func runSerializeSuiteMetadataFlow(ctx context.Context, suite testsuite.TestSuite, conn *grpc.ClientConn) error {
-	allTestMetadata := map[string]*core_api_bindings.TestMetadata{}
-	for testName, test := range suite.GetTests() {
-		testConfigBuilder := testsuite.NewTestConfigurationBuilder()
-		test.Configure(testConfigBuilder)
-		testConfig := testConfigBuilder.Build()
-		usedArtifactUrls := map[string]bool{}
-		for _, artifactUrl := range testConfig.FilesArtifactUrls {
-			usedArtifactUrls[artifactUrl] = true
-		}
-		testMetadata := &core_api_bindings.TestMetadata{
-			IsPartitioningEnabled: testConfig.IsPartitioningEnabled,
-			UsedArtifactUrls:      usedArtifactUrls,
-			TestSetupTimeoutInSeconds: testConfig.SetupTimeoutSeconds,
-			TestRunTimeoutInSeconds: testConfig.RunTimeoutSeconds,
-		}
-		allTestMetadata[testName] = testMetadata
-	}
+	grpcServerResultChan := make(chan error)
 
-	networkWidthBits := suite.GetNetworkWidthBits()
-	testSuiteMetadata := &core_api_bindings.TestSuiteMetadata{
-		TestMetadata:     allTestMetadata,
-		NetworkWidthBits: networkWidthBits,
-	}
+	go func() {
+		var resultErr error = nil
+		if err := grpcServer.Serve(listener); err != nil {
+			resultErr = stacktrace.Propagate(err, "The gRPC server exited with an error")
+		}
+		grpcServerResultChan <- resultErr
+	}()
 
-	metadataSerializationClient := core_api_bindings.NewSuiteMetadataSerializationServiceClient(conn)
-	if _, err := metadataSerializationClient.SerializeSuiteMetadata(ctx, testSuiteMetadata); err != nil {
-		return stacktrace.Propagate(err, "An error occurred sending the suite metadata to the Kurtosis API server")
+	// Wait until we get a shutdown signal
+	<- termSignalChan
+
+	serverStoppedChan := make(chan interface{})
+	go func() {
+		grpcServer.GracefulStop()
+		serverStoppedChan <- nil
+	}()
+	select {
+	case <- serverStoppedChan:
+		logrus.Info("gRPC server has exited gracefully")
+	case <- time.After(grpcServerStopGracePeriod):
+		logrus.Warnf("gRPC server failed to stop gracefully after %v; hard-stopping now...", grpcServerStopGracePeriod)
+		grpcServer.Stop()
+		logrus.Info("gRPC server was forcefully stopped")
+	}
+	if err := <- grpcServerResultChan; err != nil {
+		// Technically this doesn't need to be an error, but we make it so to fail loudly
+		return stacktrace.Propagate(err, "gRPC server returned an error after it was done serving")
 	}
 
 	return nil
+
 }
-
-func runTestExecutionFlow(ctx context.Context, suite testsuite.TestSuite, conn *grpc.ClientConn) error {
-	executionClient := core_api_bindings.NewTestExecutionServiceClient(conn)
-	testExecutionInfo, err := executionClient.GetTestExecutionInfo(ctx, &emptypb.Empty{})
-	if err != nil {
-		return stacktrace.Propagate(err, "An error occurred getting the test execution info")
-	}
-	testName := testExecutionInfo.TestName
-
-	allTests := suite.GetTests()
-	test, found := allTests[testName]
-	if !found {
-		return stacktrace.NewError(
-			"Testsuite was directed to execute test '%v', but no test with that name exists " +
-				"in the testsuite; this is a Kurtosis code bug",
-			testName)
-	}
-
-	testConfigBuilder := testsuite.NewTestConfigurationBuilder()
-	test.Configure(testConfigBuilder)
-	testConfig := testConfigBuilder.Build()
-	filesArtifactUrls := testConfig.FilesArtifactUrls
-
-	networkCtx := networks.NewNetworkContext(
-		executionClient,
-		filesArtifactUrls)
-
-	logrus.Info("Setting up the test network...")
-	// Kick off a timer with the API in case there's an infinite loop in the user code that causes the test to hang forever
-	logrus.Debug("Registering test setup with API container...")
-	if _, err := executionClient.RegisterTestSetup(ctx, &emptypb.Empty{}); err != nil {
-		return stacktrace.Propagate(err, "An error occurred registering the test setup with the API container")
-	}
-	logrus.Debug("Test setup registered with API container")
-	logrus.Debug("Executing setup logic...")
-	untypedNetwork, err := test.Setup(networkCtx)
-	if err != nil {
-		return stacktrace.Propagate(err, "An error occurred setting up the test network")
-	}
-	logrus.Debug("Setup logic executed")
-	logrus.Debug("Registering test setup completion...")
-	if _, err := executionClient.RegisterTestSetupCompletion(ctx, &emptypb.Empty{}); err != nil {
-		return stacktrace.Propagate(err, "An error occurred registering the test setup completion with the API container")
-	}
-	logrus.Debug("Test setup completion registered")
-	logrus.Info("Test network set up")
-
-	logrus.Infof("Executing test '%v'...", testName)
-	if _, err := executionClient.RegisterTestExecution(ctx, &emptypb.Empty{}); err != nil {
-		return stacktrace.Propagate(err, "An error occurred registering the test execution with the API container")
-	}
-	testResultErr := runTest(test, untypedNetwork)
-	logrus.Tracef("After running test: resultErr: %v", testResultErr)
-	logrus.Infof("Executed test '%v'", testName)
-
-	if testResultErr != nil {
-		return stacktrace.Propagate(testResultErr, "An error occurred when running the test")
-	}
-
-	return nil
-}
-