+++ golang/lib/rpc_api/rpc_api_consts/rpc_api_consts.go
@@ -1,0 +1,6 @@
+package rpc_api_consts
+
+const (
+	ListenProtocol = "tcp"
+	ListenPort = 7718
+)